Get the device-read node up and running
Added a new node which just accepts the deviceID and the building ID and fetches for that device the current state.

This is needed in order to check if the device was manipulated manually between two automatic commands and to prevent overwriting manual commands by automatic ones.

This will close #18